refactor: renames table component file
